go schemabuilder/input: Move Input/Args functions to separate file.
Summary: Moves all the schemabuilder functions that are related to
building input parameter parsing into it's own file.  In the process we
also reordered some of the functions to couple logic together.
