+++ graphql/schemabuilder/input.go
@@ -1,0 +1,419 @@
+package schemabuilder
+
+import (
+	"encoding/base64"
+	"errors"
+	"fmt"
+	"reflect"
+	"time"
+
+	"github.com/samsarahq/thunder/graphql"
+	"github.com/samsarahq/thunder/internal"
+)
+
+type argField struct {
+	field    reflect.StructField
+	parser   *argParser
+	optional bool
+}
+
+type argParser struct {
+	FromJSON func(interface{}, reflect.Value) error
+	Type     reflect.Type
+}
+
+func (p *argParser) Parse(args interface{}) (interface{}, error) {
+	if p == nil {
+		return nilParseArguments(args)
+	}
+	parsed := reflect.New(p.Type).Elem()
+	if err := p.FromJSON(args, parsed); err != nil {
+		return nil, err
+	}
+	return parsed.Interface(), nil
+}
+
+func nilParseArguments(args interface{}) (interface{}, error) {
+	if args == nil {
+		return nil, nil
+	}
+	if args, ok := args.(map[string]interface{}); !ok || len(args) != 0 {
+		return nil, graphql.NewSafeError("unexpected args")
+	}
+	return nil, nil
+}
+
+func (sb *schemaBuilder) makeStructParser(typ reflect.Type) (*argParser, graphql.Type, error) {
+	argType, fields, err := sb.getStructObjectFields(typ)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return &argParser{
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asMap, ok := value.(map[string]interface{})
+			if !ok {
+				return errors.New("not an object")
+			}
+
+			for name, field := range fields {
+				value := asMap[name]
+				fieldDest := dest.FieldByIndex(field.field.Index)
+				if err := field.parser.FromJSON(value, fieldDest); err != nil {
+					return fmt.Errorf("%s: %s", name, err)
+				}
+			}
+			for name := range asMap {
+				if _, ok := fields[name]; !ok {
+					return fmt.Errorf("unknown arg %s", name)
+				}
+			}
+
+			return nil
+		},
+		Type: typ,
+	}, argType, nil
+}
+
+func (sb *schemaBuilder) getStructObjectFields(typ reflect.Type) (*graphql.InputObject, map[string]argField, error) {
+	// Check if the struct type is already cached
+	if cached, ok := sb.typeCache[typ]; ok {
+		return cached.argType, cached.fields, nil
+	}
+
+	fields := make(map[string]argField)
+	argType := &graphql.InputObject{
+		Name:        typ.Name(),
+		InputFields: make(map[string]graphql.Type),
+	}
+	if argType.Name != "" {
+		argType.Name += "_InputObject"
+	}
+
+	if typ.Kind() != reflect.Struct {
+		return nil, nil, fmt.Errorf("expected struct but received type %s", typ.Name())
+	}
+
+	// Cache type information ahead of time to catch self-reference
+	sb.typeCache[typ] = cachedType{argType, fields}
+
+	for i := 0; i < typ.NumField(); i++ {
+		field := typ.Field(i)
+		if field.Anonymous {
+			return nil, nil, fmt.Errorf("bad arg type %s: anonymous fields not supported", typ)
+		}
+
+		fieldInfo, err := parseGraphQLFieldInfo(field)
+		if err != nil {
+			return nil, nil, fmt.Errorf("bad type %s: %s", typ, err.Error())
+		}
+		if fieldInfo.Skipped {
+			continue
+		}
+
+		if _, ok := fields[fieldInfo.Name]; ok {
+			return nil, nil, fmt.Errorf("bad arg type %s: duplicate field %s", typ, fieldInfo.Name)
+		}
+		parser, fieldArgTyp, err := sb.makeArgParser(field.Type)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		fields[fieldInfo.Name] = argField{
+			field:  field,
+			parser: parser,
+		}
+		argType.InputFields[fieldInfo.Name] = fieldArgTyp
+	}
+
+	return argType, fields, nil
+}
+
+func (sb *schemaBuilder) makeArgParser(typ reflect.Type) (*argParser, graphql.Type, error) {
+	if typ.Kind() == reflect.Ptr {
+		parser, argType, err := sb.makeArgParserInner(typ.Elem())
+		if err != nil {
+			return nil, nil, err
+		}
+		return wrapPtrParser(parser), argType, nil
+	}
+
+	parser, argType, err := sb.makeArgParserInner(typ)
+	if err != nil {
+		return nil, nil, err
+	}
+	return parser, &graphql.NonNull{Type: argType}, nil
+}
+
+func (sb *schemaBuilder) makeArgParserInner(typ reflect.Type) (*argParser, graphql.Type, error) {
+	if sb.enumMappings[typ] != nil {
+		parser, argType := sb.getEnumArgParser(typ)
+		return parser, argType, nil
+	}
+
+	if parser, argType, ok := getScalarArgParser(typ); ok {
+		return parser, argType, nil
+	}
+
+	switch typ.Kind() {
+	case reflect.Struct:
+		parser, argType, err := sb.makeStructParser(typ)
+		if err != nil {
+			return nil, nil, err
+		}
+		if argType.(*graphql.InputObject).Name == "" {
+			return nil, nil, fmt.Errorf("bad type %s: should have a name", typ)
+		}
+		return parser, argType, nil
+	case reflect.Slice:
+		return sb.makeSliceParser(typ)
+	default:
+		return nil, nil, fmt.Errorf("bad arg type %s: should be struct, scalar, pointer, or a slice", typ)
+	}
+}
+
+func wrapPtrParser(inner *argParser) *argParser {
+	return &argParser{
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			if value == nil {
+				// optional value
+				return nil
+			}
+
+			ptr := reflect.New(inner.Type)
+			if err := inner.FromJSON(value, ptr.Elem()); err != nil {
+				return err
+			}
+			dest.Set(ptr)
+			return nil
+		},
+		Type: reflect.PtrTo(inner.Type),
+	}
+}
+
+func (sb *schemaBuilder) getEnumArgParser(typ reflect.Type) (*argParser, graphql.Type) {
+	var values []string
+	for mapping := range sb.enumMappings[typ].Map {
+		values = append(values, mapping)
+	}
+	return &argParser{FromJSON: func(value interface{}, dest reflect.Value) error {
+		asString, ok := value.(string)
+		if !ok {
+			return errors.New("not a string")
+		}
+		val, ok := sb.enumMappings[typ].Map[asString]
+		if !ok {
+			return fmt.Errorf("unknown enum value %v", asString)
+		}
+		dest.Set(reflect.ValueOf(val).Convert(dest.Type()))
+		return nil
+	}, Type: typ}, &graphql.Enum{Type: typ.Name(), Values: values, ReverseMap: sb.enumMappings[typ].ReverseMap}
+
+}
+
+func (sb *schemaBuilder) makeSliceParser(typ reflect.Type) (*argParser, graphql.Type, error) {
+	inner, argType, err := sb.makeArgParser(typ.Elem())
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return &argParser{
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asSlice, ok := value.([]interface{})
+			if !ok {
+				return errors.New("not a list")
+			}
+
+			dest.Set(reflect.MakeSlice(typ, len(asSlice), len(asSlice)))
+
+			for i, value := range asSlice {
+				if err := inner.FromJSON(value, dest.Index(i)); err != nil {
+					return err
+				}
+			}
+
+			return nil
+		},
+		Type: typ,
+	}, &graphql.List{Type: argType}, nil
+}
+
+func getScalarArgParser(typ reflect.Type) (*argParser, graphql.Type, bool) {
+	for match, argParser := range scalarArgParsers {
+		if internal.TypesIdenticalOrScalarAliases(match, typ) {
+			name, ok := getScalar(typ)
+			if !ok {
+				panic(typ)
+			}
+
+			if typ != argParser.Type {
+				// The scalar may be a type alias here,
+				// so we annotate the parser to output the
+				// alias instead of the underlying type.
+				newParser := *argParser
+				newParser.Type = typ
+				argParser = &newParser
+			}
+
+			return argParser, &graphql.Scalar{Type: name}, true
+		}
+	}
+	return nil, nil, false
+}
+
+var scalarArgParsers = map[reflect.Type]*argParser{
+	reflect.TypeOf(bool(false)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asBool, ok := value.(bool)
+			if !ok {
+				return errors.New("not a bool")
+			}
+			dest.Set(reflect.ValueOf(asBool).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(float64(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(asFloat).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(float32(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(float32(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(int64(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(int64(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(int32(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(int32(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(int16(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(int16(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(int8(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(int8(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(uint64(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(int64(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(uint32(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(uint32(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(uint16(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(uint16(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(uint8(0)): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asFloat, ok := value.(float64)
+			if !ok {
+				return errors.New("not a number")
+			}
+			dest.Set(reflect.ValueOf(uint8(asFloat)).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(string("")): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asString, ok := value.(string)
+			if !ok {
+				return errors.New("not a string")
+			}
+			dest.Set(reflect.ValueOf(asString).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf([]byte{}): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asString, ok := value.(string)
+			if !ok {
+				return errors.New("not a string")
+			}
+			bytes, err := base64.StdEncoding.DecodeString(asString)
+			if err != nil {
+				return err
+			}
+			dest.Set(reflect.ValueOf(bytes).Convert(dest.Type()))
+			return nil
+		},
+	},
+	reflect.TypeOf(time.Time{}): {
+		FromJSON: func(value interface{}, dest reflect.Value) error {
+			asString, ok := value.(string)
+			if !ok {
+				return errors.New("not a string")
+			}
+			asTime, err := time.Parse(time.RFC3339, asString)
+			if err != nil {
+				return errors.New("not an iso8601 time")
+			}
+			dest.Set(reflect.ValueOf(asTime).Convert(dest.Type()))
+			return nil
+		},
+	},
+}
+
+func init() {
+	for typ, arg := range scalarArgParsers {
+		arg.Type = typ
+	}
+}
